# api-client/src/generated/api.schemas.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * AI-Solutions API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Solution {
  id: number;
  title: string;
  description: string;
  category: string;
  icon: string;
  features: string[];
  /** @nullable */
  imageUrl?: string | null;
}

export interface Industry {
  id: number;
  name: string;
  sector: string;
  description: string;
  outcome: string;
  year: number;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface Testimonial {
  id: number;
  clientName: string;
  company: string;
  role: string;
  rating: number;
  message: string;
  /** @nullable */
  avatarUrl: string | null;
}

export interface Article {
  id: number;
  title: string;
  summary: string;
  content: string;
  author: string;
  publishedAt: string;
  category: string;
  /** @nullable */
  imageUrl: string | null;
}

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  location: string;
  type: string;
  isUpcoming?: boolean;
  photos: string[];
  /** @nullable */
  coverImageUrl?: string | null;
}

export interface InquiryInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone: string;
  companyName: string;
  country: string;
  jobTitle: string;
  jobDetails: string;
}

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  phone: string;
  companyName: string;
  country: string;
  jobTitle: string;
  jobDetails: string;
  createdAt: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminToken {
  token: string;
}

export interface CountGroup {
  label: string;
  count: number;
}

export interface AdminStats {
  totalInquiries: number;
  inquiriesByCountry: CountGroup[];
  inquiriesByJobTitle: CountGroup[];
  recentInquiries: Inquiry[];
}

export type DeleteInquiry200 = {
  success?: boolean;
  id?: number;
};