# api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * AI-Solutions API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CountGroup } from './countGroup';
import type { Inquiry } from './inquiry';

export interface AdminStats {
  totalInquiries: number;
  inquiriesByCountry: CountGroup[];
  inquiriesByJobTitle: CountGroup[];
  recentInquiries: Inquiry[];
}