# api-zod/src/generated/types/article.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * AI-Solutions API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Article {
  id: number;
  title: string;
  summary: string;
  content: string;
  author: string;
  publishedAt: string;
  category: string;
  /** @nullable */
  imageUrl: string | null;
}