# api-zod/src/generated/types/countGroup.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * AI-Solutions API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CountGroup {
  label: string;
  count: number;
}