# api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * AI-Solutions API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: number;
  title: string;
  description: string;
  date: string;
  location: string;
  type: string;
  isUpcoming?: boolean;
  photos: string[];
  /** @nullable */
  coverImageUrl?: string | null;
}