# api-zod/src/generated/types/industry.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * AI-Solutions API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Industry {
  id: number;
  name: string;
  sector: string;
  description: string;
  outcome: string;
  year: number;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}