# api-zod/src/generated/types/inquiry.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * AI-Solutions API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Inquiry {
  id: number;
  name: string;
  email: string;
  phone: string;
  companyName: string;
  country: string;
  jobTitle: string;
  jobDetails: string;
  createdAt: string;
}