# api-zod/src/generated/types/inquiryInput.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * AI-Solutions API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InquiryInput {
  /** @minLength 1 */
  name: string;
  email: string;
  phone: string;
  companyName: string;
  country: string;
  jobTitle: string;
  jobDetails: string;
}