# api-zod/src/generated/types/solution.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * AI-Solutions API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Solution {
  id: number;
  title: string;
  description: string;
  category: string;
  icon: string;
  features: string[];
  /** @nullable */
  imageUrl?: string | null;
}