# api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Api
 * AI-Solutions API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  clientName: string;
  company: string;
  role: string;
  rating: number;
  message: string;
  /** @nullable */
  avatarUrl: string | null;
}